Update campominado.c
Finally a ready to play minefield game. Still there are some possible improvements such as:
- Prevent chose the same position previous chosen and keep the same counting plays.
- Asking if the player wants to play again. 
- This version has a game table 5x5, maybe create a possibility to choose the table size. 
But if you follow and pay attention on the game rules, it's all functional. This was a very good logical and code exercise. I will leave this way now, maybe I come back someday to new features but now, it is as it is.